requirements/run.txt
## direct dependencies
aiodns==1.1.1
aiohttp==2.2.5
alembic==0.9.9
async-timeout==2.0.1
asyncpg==0.15.0
attrs==18.1.0
git+https://github.com/BryanForbes/botus_receptus@6889b695948a1cd894d938b1d3a9a443ce21154b#egg=botus_receptus[db]
cchardet==2.1.1
click==6.7
mypy-extensions==0.3.0
psycopg2==2.7.4
SQLAlchemy==1.2.6
uvloop==0.9.1

## The following requirements were added by pip freeze:
chardet==3.0.4
idna==2.6
Mako==1.0.7
MarkupSafe==1.0
multidict==4.2.0
pycares==2.3.0
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
websockets==3.4
yarl==1.1.1
